refactor(sqlite): 2 つの RPC スタックを SqliteGateway に統合 (PBI 05)
SqliteClient.callInternal（SW → Offscreen）と
dashboardSqliteService.callDashboard（Dashboard → SW）が別々に持っていた
error 分類・result 形状を、統一 SqliteResult<T> を公開する SqliteGateway
に集約する。

- SqliteGateway: query / mutate / maintain / status の 4 メソッドと
  SqliteResult<T> = {success:true,data} | {success:false,error} を公開
- SqliteClient / dashboardSqliteService は Gateway 委譲の shim に縮小
- DashboardSqliteGateway.callDashboard: 失敗レスポンスと decode 失敗は
  SW が既に分類済みの文言をそのまま透過（二重 categorizeError を回避）。
  transport 例外のみ分類する
- queryPlan.ts に buildExtraWhereSql を集約し、IdbVfsBackend /
  opfsWorker/searchHandlers の個別実装を削除
- StorageBackend を Queryable / Mutable の 2 facet に分割
- buildExtraWhereSql の引数型から exactOptionalPropertyTypes と衝突する
  & Record<string, unknown> を除去
